requirements.txt
python-telegram-bot>=21.0
openai>=1.0.0
faiss-cpu>=1.8.0
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
numpy>=1.26.0
